pyproject.toml
[project]
name = "sirios-mosaic"
version = "0.1.0"
description = "Sirios Mosaic - NL→Views modular architecture (FastAPI)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.1",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.8.0",
    "pyyaml>=6.0.1",
    "psycopg[binary]>=3.2.1",
    "python-dateutil>=2.9.0.post0",
    "httpx>=0.27,<0.28"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]
[project.optional-dependencies]
observability = [
  "prometheus-client==0.20.0",
  "python-json-logger==2.0.7"
]
